+++ layabase/database_sqlalchemy.py
@@ -353,6 +353,7 @@ class CRUDModel:
             column.default.arg for column in sql_alchemy_field.columns if column.default
         ]
         if defaults:
+            # TODO Set marshmallow_field.default instead ?
             marshmallow_field.metadata["sqlalchemy_default"] = defaults[0]
 
         # Auto incremented field
@@ -367,18 +368,26 @@ class CRUDModel:
     @classmethod
     def query_get_parser(cls) -> reqparse.RequestParser:
         query_get_parser = cls._query_parser()
-        query_get_parser.add_argument("limit", type=inputs.positive)
-        query_get_parser.add_argument("order_by", type=str, action="append")
+        query_get_parser.add_argument("limit", type=inputs.positive, location="args")
+        query_get_parser.add_argument(
+            "order_by", type=str, action="append", location="args"
+        )
         if _supports_offset(cls.metadata.bind.url.drivername):
-            query_get_parser.add_argument("offset", type=inputs.natural)
+            query_get_parser.add_argument(
+                "offset", type=inputs.natural, location="args"
+            )
         return query_get_parser
 
     @classmethod
     def query_get_history_parser(cls) -> reqparse.RequestParser:
         query_get_hist_parser = cls._query_parser()
-        query_get_hist_parser.add_argument("limit", type=inputs.positive)
+        query_get_hist_parser.add_argument(
+            "limit", type=inputs.positive, location="args"
+        )
         if _supports_offset(cls.metadata.bind.url.drivername):
-            query_get_hist_parser.add_argument("offset", type=inputs.natural)
+            query_get_hist_parser.add_argument(
+                "offset", type=inputs.natural, location="args"
+            )
         return query_get_hist_parser
 
     @classmethod
@@ -395,9 +404,10 @@ class CRUDModel:
         for marshmallow_field in cls.schema().fields.values():
             query_parser.add_argument(
                 marshmallow_field.name,
-                required=False,
+                required=marshmallow_field.metadata.get("required_on_query", False),
                 type=_get_python_type(marshmallow_field),
                 action="append",
+                location="args",
             )
         return query_parser